refactor: Make role assignment optional when creating certs  (#409)
## Description

<!--Why this PR? What is changed? What is the effect? etc.-->

Updating the `deployment-scripts/create-kv-certificate` to add a param
that controls the running of the role assignment operation defined in
the `rbacKv` resource block.

This is helpful when using an existing Managed Identity that already has
the necessary role assigned to it. Without this param, deploys fail with
the "role assignment already exists error".

## Updating an existing module

<!--Run through the checklist if your PR updates an existing module.-->

- [ ] This is a bug fix:
- [ ] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [ ] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [x] I have run `brm validate` locally to verify the module files.
- [ ] I have run deployment tests locally to ensure the module is
deployable.
- [x] I have read the [Updating an existing
module](https://github.com/Azure/bicep-registry-modules/blob/main/CONTRIBUTING.md#updating-an-existing-module)
section in the contributing guide and updated the `version.json` file
properly:
- [ ] The PR contains backwards compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`.
- [x] The PR contains backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] The PR contains breaking changes, and I have bumped the MAJOR
version in `version.json`.
- [x] I have updated the examples in README with the latest module
version number.