fix: enhance zone redundancy configuration and fix issue with deployment to westus (#446)
## Description

Closes #445 

Changes:
- Parameterises zone redundancy configurations, so users of the module
can config to their use case.
- Default number of required zones to 3, so that zone redundancy support
is only enabled in a region with 'full' availability zone support. User
can override via the new parameters if desired and ultimately disable
the functionality if not required via `zoneRedundancyEnabled`.
- shardCount is currently enforced to be 1 or more when using Premium
SKU however some users may wish to use a Premium SKU without enabling
clustering mode. updated allow shardCount to be zero and nullify the
shardCount attribute if set zero.

**Considering this as a breaking change, due to the change in default
for `numberOfZones` from `1` to `3` and the addition of the
`zoneRedundancyEnabled` parameter for disabling this in regions that do
not support AZs fully.**

I have tested locally integrated with out own internal template that
consumes this module, with its default values against `eastus` and with
`zoneRedundancyEnabled` as false in `westus`. It worked as expected,
`eastus` enabled zone redundancy across all three zones and in `westus`
no zone redundancy was enabled (fixing #445).

I have re-run the same test with `shardCount` zero to validate that this
successfully disables this feature as intended.

## Updating an existing module

<!--Run through the checklist if your PR updates an existing module.-->

- [x] This is a bug fix:
- [x] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [ ] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [x] I have run `brm validate` locally to verify the module files.
- [x] I have run deployment tests locally to ensure the module is
deployable.
- [x] I have read the [Updating an existing
module](https://github.com/Azure/bicep-registry-modules/blob/main/CONTRIBUTING.md#updating-an-existing-module)
section in the contributing guide and updated the `version.json` file
properly:
- [ ] The PR contains backwards compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`.
- [ ] The PR contains backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [x] The PR contains breaking changes, and I have bumped the MAJOR
version in `version.json`.
- [ ] I have updated the examples in README with the latest module
version number.