feat: Added workflow for syncing CSV labels to GitHub Labels (#1285)
**Test Run:**
https://github.com/jtracey93/bicep-registry-modules/actions/runs/8285523952/job/22673465780

This pull request introduces a new GitHub Actions workflow in the
`.github/workflows/avm.platform.sync-repo-labels-from-csv.yml` file. The
workflow is designed to synchronize labels from a CSV file to GitHub
labels. It is scheduled to run daily at 3:45 AM PST and can also be
manually triggered. The workflow runs on an Ubuntu-latest environment
and uses a PowerShell script to perform the synchronization.

Key changes include:

* New GitHub Actions workflow: A new workflow named
`avm.platform.sync-repo-labels-from-csv` has been added. This workflow
is set to run on a schedule and can also be manually triggered.
* Concurrency settings: The workflow is configured to allow only one
concurrent deployment. If a new deployment is triggered while another
one is in progress, the in-progress deployment will be cancelled.
* Permissions: The workflow has write permissions for issues and pull
requests.
* Workflow job: The workflow contains a job named `sync-labels` that
runs on an `ubuntu-latest` environment. The job uses a PowerShell script
to synchronize labels from a CSV file to GitHub labels.