feat: Convert publishModuleIndex to PWSH (#1533)
## Description


[AB#33918](https://dev.azure.com/CSUSolEng/Azure%20Verified%20Modules/_workitems/edit/33918)

This pull request includes changes to the bicep module index JSON file
generation process. The primary changes include the introduction of a
new GitHub workflow and PowerShell script to generate the module index
and the removal of the previous workflow and JavaScript script. The new
workflow is scheduled to run daily at 3:45 AM PST and uses Azure login
and PowerShell to generate and upload the module index to Azure blob
storage. The new PowerShell script generates the module index based on
the modules in the repository and merges it with the previous version of
the index.

> Diffs have been compared and apart from a move of
`avm/res/app-configuration/configuration-store` in the JSON file as its
being sorted differently, there are no regressions. You can check for
yourself in VScode by downloading, saving and comparing the 2 files
from: [current
workflow](https://biceplivedatasaprod.blob.core.windows.net/bicep-cdn-live-data-container/module-index)
VS [new
workflow](https://stguksjtbrmindextest.blob.core.windows.net/bicep-cdn-live-data-container/module-index)

Changes to GitHub workflows:

*
[`.github/workflows/avm.platform.publish-module-index-json.yml`](diffhunk://#diff-a32f4d20d0b1b49f6307c794b2b75c9515a36131eef33ccab3012a2c3c431ae3R1-R77):
Added a new GitHub workflow that is scheduled to run daily at 3:45 AM
PST. This workflow checks out the repository, logs in to Azure, installs
Azure PowerShell Modules, generates the module index, uploads the
artifacts, and uploads the module index to Azure blob storage.
*
[`.github/workflows/publish-module-index.yml`](diffhunk://#diff-0db9403318bc32b133eb4c1f5af094a003d067d3cce1e5839b773910d8536cb0L1-L118):
Removed the previous GitHub workflow that was used to publish the module
index.

Changes to scripts:

*
[`avm/utilities/pipelines/platform/Invoke-AvmJsonModuleIndexGeneration.ps1`](diffhunk://#diff-d93730e5546a51dc8ca6236af545822fff02a6b4b551eb6d0a34ac4e457446a1R1-R169):
Added a new PowerShell script that generates the module index based on
the modules in the repository and merges it with the previous version of
the index. This script is used in the new GitHub workflow.
*
[`scripts/github-actions/generate-module-index-data.js`](diffhunk://#diff-2c75227ba0b424f38faa9fe12c0fe64333ce433bbca33986e000139d8445c399L1-L211):
Removed the previous JavaScript script that was used to generate the
module index.

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->

| Pipeline |
| -------- |
|
[![avm.platform.publish-module-index-json](https://github.com/jtracey93/bicep-registry-modules/actions/workflows/avm.platform.publish-module-index-json.yml/badge.svg)](https://github.com/jtracey93/bicep-registry-modules/actions/workflows/avm.platform.publish-module-index-json.yml)
|

> Had to merge in main to resolve a conflict which wiped the workflow
from my fork. However, last run can be seen here:
https://github.com/jtracey93/bicep-registry-modules/actions/runs/8572556792/job/23495383080
and comparison of files above in storage accounts is sufficient evidence

## Type of Change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] Update to CI Environment or utlities (Non-module effecting
changes)
- [ ] Azure Verified Module updates:
- [ ] Bugfix containing backwards compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [ ] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [ ] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [x] My corresponding pipelines / checks run clean and green without
any errors or warnings

<!-- Please keep up to day with the contribution guide at
https://aka.ms/avm/contribute/bicep -->

---------

Co-authored-by: Alexander Sehr <ASehr@hotmail.de>