feat: Update container group main.json and test deployment
The container group main.json file was updated to use a newer version of the bicep generator (0.27.1.19265) and a different template hash. Additionally, the livenessProbe property was added to the container group definition.

In the test deployment file, the commented out environment variables were removed and the CLIENT_ID and CLIENT_SECRET environment variables were added with test values.