feat: update logic to create a meaningful issue, if the module list template is out of sync with the CSV files (#1918)
If the template file is not in sync with the CSV files, an issue was
created. The old issue was not very helpful, thus the new issue will
show exactly what it is about and shows in detail the diff. Here you can
see how such an issue will look like:
https://github.com/Azure/GHPolicyTest/issues/189