feat: Add support for Azure Container Instances in `avm/ptn/dev-ops/cicd-agents-and-runners`
This commit adds support for Azure Container Instances in the `avm/ptn/dev-ops/cicd-agents-and-runners` module. It includes the necessary code changes to specify the existing network container instance subnet name and the existing network container app subnet name. These changes are required for the deployment script to run the placeholder agent and the container app in the correct subnets. The subnets also need to have service delegation for Azure Container Instances and App environments respectively.

Co-authored-by: Jack Tracey <41163455+jtracey93@users.noreply.github.com>
Co-authored-by: Erika Gressi <56914614+eriqua@users.noreply.github.com>
Co-authored-by: Felipe Binotto <felipe_binotto@hotmail.com>