feat: Update `avm/ptn/dev-ops/cicd-agents-and-runners` module
This commit adds support for the `avm/ptn/dev-ops/cicd-agents-and-runners` module. It includes the necessary code changes to specify the existing network container instance subnet name and the existing network container app subnet name. These changes are required for the deployment script to run the placeholder agent and the container app in the correct subnets. The subnets also need to have service delegation for Azure Container Instances and App environments respectively.