feat: Update `avm/ptn/dev-ops/cicd-agents-and-runners` module
This commit updates the `avm/ptn/dev-ops/cicd-agents-and-runners` module. It includes code changes to set the `zoneRedundant` property to `true` and the `internal` property to `true` in the `main.bicep` file. These changes ensure that the module supports zone redundancy and internal networking when `privateNetworking` is enabled. The commit also updates the `main.json` file to change the `templateHash` value.

Co-authored-by: Jack Tracey <41163455+jtracey93@users.noreply.github.com>
Co-authored-by: Erika Gressi <56914614+eriqua@users.noreply.github.com>
Co-authored-by: Felipe Binotto <felipe_binotto@hotmail.com>