fix: Update storage account template URLs
Updated the template URLs for `Microsoft.Storage/storageAccounts/fileServices` and `Microsoft.Storage/storageAccounts/queueServices` to use the correct version, [2023-04-01](https://learn.microsoft.com/en-us/azure/templates/Microsoft.Storage/2023-04-01/storageAccounts/fileServices) and [2023-04-01](https://learn.microsoft.com/en-us/azure/templates/Microsoft.Storage/2023-04-01/storageAccounts/queueServices) respectively. This ensures that the correct versions are used for these services.

Refactor: Update bicep generator version

Updated the version of the bicep generator used in the `main.json` file to `0.30.23.60470`. This ensures that the latest version of the generator is used for the deployment.

Refactor: Update templateHash values

Updated the `templateHash` values in the `main.json` file to reflect the changes made in the template. This ensures that the correct templateHash values are used for the deployment.

docs: Update README for `avm/ptn/ai-platform/baseline`

Updated the README file for the `avm/ptn/ai-platform/baseline` module. This update provides more accurate and up-to-date information about the module.