feat: Add strong types to `avm/res/sql/server` and it's children  (#3582)
## Description

This PR is aiming to add strong types to parameters for
`avm/res/sq/server` and it's children.

This PR contains typing for the `server`, `database` and `elastic-pool`.
Other modules will be tackled later, as I'd like to see first that this
is a good direction and approach adding these.

* Types are exported from children and imported into the main module
* To align with the latest API there is an SKU object for databases and
elastic pools, instead of individual `skuName`, `skuTier` etc.
* Default values are provided in the child module, and the parent is not
duplicating those default value assignment
* Also I have checked the latest API, and turned out that some of the
new parameters were missing, so I went ahead and added them to have a
complete coverage
* Updated `minCapacity` to enable fractional values both at database and
elastic-pool modules
* As a consequence version number has been increased
* Updated tests where it was necessary. All tests are successfully
completed

### Questions

* What is the best way to expose the types of the parameters of the
child module with the least repetition? Currently I need to define a
given module parameter for the module (with it's description), and then
add the same to the type again with the same description. Is there a
better DRY way to do this?
* Also is there a better way to assign the values in the parent module
to the parameters to the child module? Currently every parameter is
assigned individually, maybe there is a better and shorter way to do
that?

Contributes to #1141

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->

| Pipeline |
| -------- |
|
[![avm.res.sql.server](https://github.com/peterbud/bicep-registry-modules/actions/workflows/avm.res.sql.server.yml/badge.svg)](https://github.com/peterbud/bicep-registry-modules/actions/workflows/avm.res.sql.server.yml)
|

Pls note that in my subscription Maintenance Window feature is not
available, hence the WAF test is failing

## Type of Change

<!-- Use the checkboxes [x] on the options that are relevant. -->

- [ ] Update to CI Environment or utilities (Non-module affecting
changes)
- [ ] Azure Verified Module updates:
- [ ] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [x] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [x] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [x] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [x] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [x] My corresponding pipelines / checks run clean and green without
any errors or warnings

<!-- Please keep up to date with the contribution guide at
https://aka.ms/avm/contribute/bicep -->