fix: avm/res/storage-account - skip azureFilesIdentityBasedAuthentication when empty instead of null
When deploying this template for an existing storage account, it removes the azureFilesIdentityBasedAuthentication method when azureFilesIdentityBasedAuthentication is empty. In case of AD join, where you would use the powershell module to join the storage account, the storage acocunt loses authentication method for the share.

With this fix it skips the property all together when empty.