fix: AutomationAccount - Restricted test locations to only one (#4569)
## Description

Accounting for issue that has plagued the module for many weeks now.

Turns out, the module was not allowed to install any more modules, as it
was blocked due to the criteria `More than 25 PUT /Account calls with
201 response from same sub within the 4 hour period in 4 distinct
regions.`
-> Do not deploy the module into more than 4 regions AND do not deploy
the module more that 25 times per region within a 4 hour window
-> Do not install more than 150 modules per region AND not in more than
9 regions within a 1 hour window

Reason:
_Azure Automation continuously monitors service usage to ensure fair
access to cloud resources for all users. These measures are aimed
towards enhancing service reliability and performance while safeguarding
against misconfigurations that could adversely impact our valuable
customers. If the system detects an unusual usage pattern, it
temporarily blocks the Automation account to prevent misuse._

This restriction only applies to the type of subscription / agreement
AVM is using and was hence not an issue for our contributors.

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->

| Pipeline |
| -------- |
|
[![avm.res.automation.automation-account](https://github.com/Azure/bicep-registry-modules/actions/workflows/avm.res.automation.automation-account.yml/badge.svg?branch=users%2Falsehr%2FautoAccountSoloLoc&event=workflow_dispatch)](https://github.com/Azure/bicep-registry-modules/actions/workflows/avm.res.automation.automation-account.yml)
|

## Type of Change

<!-- Use the checkboxes [x] on the options that are relevant. -->

- [ ] Update to CI Environment or utilities (Non-module affecting
changes)
- [ ] Azure Verified Module updates:
- [x] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [ ] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation