fix: allow any string for powershell and python module versions not just latest
This was missed as part of the original PR to add support for powershell
7.2 and python 2/3 modules. The type for both was hard coded to only allow
'latest' or nothing as the value for the version of a module, but the
bicep child module for those has a default value of 'latest' as well.

This change swaps to using string as the allowed type of the version property
for both module types because the default value for the child module will
handle not providing one and allows users to specify other versions for
the modules they want to install.