fix: allow any string for powershell and python module versions not just latest in automation-account (#4649)
## Description

This was missed as part of the original PR (#4352) to add support for
powershell 7.2 and python 2/3 modules. The type for both was hard coded
to only allow 'latest' or nothing as the value for the version of a
module, but the bicep child module for those has a default value of
'latest' as well.

This change swaps to using string as the allowed type of the version
property for both module types because the default value for the child
module will handle not providing one and allows users to specify other
versions for the modules they want to install.

## Pipeline Reference

<!-- Insert your Pipeline Status Badge below -->

| Pipeline |
| -------- |
|
[![avm.res.automation.automation-account](https://github.com/chrislgardner/bicep-registry-modules/actions/workflows/avm.res.automation.automation-account.yml/badge.svg?branch=autaccount-issue2485)](https://github.com/chrislgardner/bicep-registry-modules/actions/workflows/avm.res.automation.automation-account.yml)
|

## Type of Change

<!-- Use the checkboxes [x] on the options that are relevant. -->

- [ ] Update to CI Environment or utilities (Non-module affecting
changes)
- [x] Azure Verified Module updates:
- [x] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [ ] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [x] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [x] My corresponding pipelines / checks run clean and green without
any errors or warnings

<!-- Please keep up to date with the contribution guide at
https://aka.ms/avm/contribute/bicep -->