feat: Update `avm/res/document-db/mongo-cluster` module to support Microsoft Entra `user` entities (#5260)
## Description

I'm adding a new `user` submodule to `avm/res/document-db/mongo-cluster`
so we can build Azure Developer CLI templates that support Microsoft
Entra authentication end-to-end.

Here's the relevant documentation:
<https://learn.microsoft.com/azure/cosmos-db/mongodb/vcore/how-to-configure-entra-authentication>

Unfortunately, this module was using an older RP, so this caused a few
"domino" changes:

- The `api-version` needed to be updated to at least
`2025-04-01-preview` to support Microsoft Entra authentication
  - The `nodeType` property no longer exists in the last API version
- The `highAvailabilityMode` property is now a string with multiple
allowed values

And in keeping with best practices, I just made these minor changes:

- I updated most unit tests to deploy minimal single-node resources to
minimize the amount of churn and extra resources deployed for each CI
run
- I updated the WAF-aligned unit test to deploy triplicate nodes that
are zone-redundant (no redundancy in the initial version)
- I updated the metadata to correctly reference *"Azure Cosmos DB for
MongoDB (vCore)"*
- The WAF-aligned unit test needed tags to suppress a warning from
PSRule

I did notice this behavior, but I didn't fix it in this PR:

- The `max.*` unit test has transient failures. I couldn't pin down what
was causing it, but it was due to the test not being able to "find" the
nested dependencies. I don't have enough information to diagnose if it
was a race condition or a naming collision.

Finally, I didn't suppress the Bicep linter warning for the
`api-version`. That API version is supported in the REST API, but it's
not documented in the specs yet.

This PR depends on this other change to remove linter warnings:

- azure/azure-rest-api-specs#34580

> [!NOTE]
> I tried my best to scope my changes to the minimal necessary to add
this submodule and Entra authentication support.

## Pipeline Reference

| Pipeline |
| --- |
|
[![avm.res.document-db.mongo-cluster](https://github.com/seesharprun/bicep-registry-modules/actions/workflows/avm.res.document-db.mongo-cluster.yml/badge.svg)](https://github.com/seesharprun/bicep-registry-modules/actions/workflows/avm.res.document-db.mongo-cluster.yml)
|

## Type of Change

- [ ] Update to CI Environment or utilities (Non-module affecting
changes)
- [x] Azure Verified Module updates:
- [ ] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [ ] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [x] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [x] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [x] Update to documentation

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [x] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [x] My corresponding pipelines / checks run clean and green without
any errors or warnings

---------

Co-authored-by: Alexander Sehr <ASehr@hotmail.de>