chore: Update `avm/res/document-db/mongo-cluster` module tests to use the correct prefix (#5272)
## Description

The `serviceShort` value in the key vault e2e test for the
`avm/res/document-db/mongo-cluster` module incorrectly used the `ddda`
prefix for the serviceShort. This is the prefix for
"documentdb/database-account" and it caused some test collisions when
testing this service and other Azure Cosmos DB modules in paralle. The
prefix has been corrected to `ddmc` for "documentdb/mongo-cluster" to
correct this issue.

## Pipeline Reference

| Pipeline |
| --- |
|
[![avm.res.document-db.mongo-cluster](https://github.com/seesharprun/bicep-registry-modules/actions/workflows/avm.res.document-db.mongo-cluster.yml/badge.svg)](https://github.com/seesharprun/bicep-registry-modules/actions/workflows/avm.res.document-db.mongo-cluster.yml)
|

## Type of Change

- [ ] Update to CI Environment or utilities (Non-module affecting
changes)
- [ ] Azure Verified Module updates:
- [x] Bugfix containing backwards-compatible bug fixes, and I have NOT
bumped the MAJOR or MINOR version in `version.json`:
- [ ] Someone has opened a bug report issue, and I have included "Closes
#{bug_report_issue_number}" in the PR description.
- [x] The bug was found by the module author, and no one has opened an
issue to report it yet.
- [ ] Feature update backwards compatible feature updates, and I have
bumped the MINOR version in `version.json`.
- [ ] Breaking changes and I have bumped the MAJOR version in
`version.json`.
  - [ ] Update to documentation

## Checklist

- [x] I'm sure there are no other open Pull Requests for the same
update/change
- [x] I have run `Set-AVMModule` locally to generate the supporting
module files.
- [ ] My corresponding pipelines / checks run clean and green without
any errors or warnings